refactor: 도메인이 외부 라이브러리에 의존하지 않도록 분리
